+++ openmls/src/group/mls_group/mod.rs
@@ -3,20 +3,15 @@
 //! This module contains [`MlsGroup`] and its submodules.
 //!
 
-use create_commit::CreateCommitParams;
 use past_secrets::MessageSecretsStore;
 use proposal_store::ProposalQueue;
 use serde::{Deserialize, Serialize};
-use staged_commit::{MemberStagedCommitState, StagedCommitState};
 use tls_codec::Serialize as _;
 
 #[cfg(test)]
 use crate::treesync::node::leaf_node::TreePosition;
 
-use super::{
-    diff::compute_path::PathComputationResult,
-    proposal_store::{ProposalStore, QueuedProposal},
-};
+use super::proposal_store::{ProposalStore, QueuedProposal};
 use crate::{
     binary_tree::array_representation::LeafNodeIndex,
     ciphersuite::{hash_ref::ProposalRef, signable::Signable},
@@ -24,16 +19,16 @@ use crate::{
     error::LibraryError,
     framing::{mls_auth_content::AuthenticatedContent, *},
     group::{
-        CreateCommitError, CreateGroupContextExtProposalError, Extension, ExtensionType,
-        Extensions, ExternalPubExtension, GroupContext, GroupEpoch, GroupId, MlsGroupJoinConfig,
-        MlsGroupStateError, OutgoingWireFormatPolicy, ProposalQueueError, PublicGroup,
-        RatchetTreeExtension, RequiredCapabilitiesExtension, StagedCommit,
+        CreateGroupContextExtProposalError, Extension, ExtensionType, Extensions,
+        ExternalPubExtension, GroupContext, GroupEpoch, GroupId, MlsGroupJoinConfig,
+        MlsGroupStateError, OutgoingWireFormatPolicy, PublicGroup, RatchetTreeExtension,
+        RequiredCapabilitiesExtension, StagedCommit,
     },
     key_packages::KeyPackageBundle,
     messages::{
         group_info::{GroupInfo, GroupInfoTBS, VerifiableGroupInfo},
         proposals::*,
-        Commit, ConfirmationTag, GroupSecrets, Welcome,
+        ConfirmationTag, GroupSecrets, Welcome,
     },
     schedule::{
         message_secrets::MessageSecrets,
@@ -54,7 +49,6 @@ use crate::schedule::{application_export_tree::ApplicationExportTree, Applicatio
 
 // Private
 mod application;
-mod creation;
 mod exporting;
 mod updates;
 
@@ -64,7 +58,7 @@ use config::*;
 pub(crate) mod builder;
 pub(crate) mod commit_builder;
 pub(crate) mod config;
-pub(crate) mod create_commit;
+pub(crate) mod creation;
 pub(crate) mod errors;
 pub(crate) mod membership;
 pub(crate) mod past_secrets;
@@ -153,7 +147,7 @@ impl From<PendingCommitState> for StagedCommit {
 ///   allows access to all of its functionality, (except merging pending commits,
 ///   see the [`MlsGroupState::PendingCommit`] for more information) and it's the
 ///   state the group starts in (except when created via
-///   [`MlsGroup::join_by_external_commit()`], see the functions documentation for
+///   [`MlsGroup::external_commit_builder()`], see the functions documentation for
 ///   more information). From this `Operational`, the group state can either
 ///   transition to [`MlsGroupState::Inactive`], when it processes a commit that
 ///   removes this client from the group, or to [`MlsGroupState::PendingCommit`],
@@ -183,12 +177,12 @@ impl From<PendingCommitState> for StagedCommit {
 ///
 ///   * A group can enter the [`PendingCommitState::External`] sub-state only as
 ///     the initial state when the group is created via
-///     [`MlsGroup::join_by_external_commit()`]. In contrast to the
+///     [`MlsGroup::external_commit_builder()`]. In contrast to the
 ///     [`PendingCommitState::Member`] `PendingCommit` state, the only possible
 ///     functionality that can be used is the [`MlsGroup::merge_pending_commit()`]
 ///     function, which merges the pending external commit and transitions the
 ///     state to [`MlsGroupState::PendingCommit`]. For more information on the
-///     external commit process, see [`MlsGroup::join_by_external_commit()`] or
+///     external commit process, see [`MlsGroup::external_commit_builder()`] or
 ///     Section 11.2.1 of the MLS specification.
 #[derive(Debug, Serialize, Deserialize)]
 #[cfg_attr(any(test, feature = "test-utils"), derive(Clone, PartialEq))]
@@ -367,7 +361,7 @@ impl MlsGroup {
     ///
     /// Note that this has no effect if the group was created through an external commit and
     /// the resulting external commit has not been merged yet. For more
-    /// information, see [`MlsGroup::join_by_external_commit()`].
+    /// information, see [`MlsGroup::external_commit_builder()`].
     ///
     /// Use with caution! This function should only be used if it is clear that
     /// the pending commit will not be used in the group. In particular, if a